pyproject.toml
[project]
name = "lgc-marl-rware"
version = "0.1.0"
description = "LGC-MARL with Progressive Depth Evolution on RWARE"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Environment
    "gymnasium>=0.29.0",
    "rware>=1.0.3",
    # RL Training
    "torch>=2.0.0",
    # LLM
    "openai>=1.0.0",
    "transformers>=4.40.0",
    # Tracking
    "wandb>=0.16.0",
    "weave>=0.50.0",
    # Utils
    "numpy<2",
    "networkx>=3.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "overcooked-ai>=1.1.0",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
local-llm = [
    "vllm>=0.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]
